+++ src/test/java/mj/chess/BoardTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static mj.chess.utils.StringUtils.appendNewLine;
+import static mj.chess.utils.StringUtil.appendNewLine;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
